+++ packages/extended/src/lib/rich-text-editor/features/tests/rte-heading.test.ts
@@ -1,11 +1,11 @@
 import { expect } from '@esm-bundle/chai';
 import { fixture, html } from '@open-wc/testing';
 import type { Editor } from '@tiptap/core';
-import { RichTextEditorComponent } from '../rich-text-editor';
-import { RichTextFeatureHeadingComponent } from './rte-heading';
+import { RichTextEditorComponent } from '../../rich-text-editor';
+import { RichTextFeatureHeadingComponent } from '../rte-heading';
 
-import '../rich-text-editor';
-import './rte-heading';
+import '../../rich-text-editor';
+import '../rte-heading';
 
 describe('RTE Heading Feature', () => {
   it('should contain shadow root', async () => {
