+++ packages/extended/src/lib/rich-text-editor/features/tests/rte-italic.test.ts
@@ -1,11 +1,11 @@
 import { expect } from '@esm-bundle/chai';
 import { fixture, html } from '@open-wc/testing';
 import type { Editor } from '@tiptap/core';
-import { RichTextEditorComponent } from '../rich-text-editor';
-import { RteItalicComponent } from './rte-italic';
+import { RichTextEditorComponent } from '../../rich-text-editor';
+import { RteItalicComponent } from '../rte-italic';
 
-import '../rich-text-editor';
-import './rte-italic';
+import '../../rich-text-editor';
+import '../rte-italic';
 
 describe('RTE Italic Feature', () => {
   it('should contain shadow root', async () => {
