+++ packages/extended/src/lib/rich-text-editor/features/tests/rte-strike.test.ts
@@ -1,46 +1,46 @@
 import { expect } from '@esm-bundle/chai';
 import { fixture, html } from '@open-wc/testing';
 import type { Editor } from '@tiptap/core';
-import { RichTextEditorComponent } from '../rich-text-editor';
-import { RichTextFeatureCodeComponent } from './rte-code';
+import { RichTextEditorComponent } from '../../rich-text-editor';
+import { RichTextFeatureStrikeComponent } from '../rte-strike';
 
-import '../rich-text-editor';
-import './rte-code';
+import '../../rich-text-editor';
+import '../rte-strike';
 
-describe('RTE Code Feature', () => {
+describe('RTE Strike Feature', () => {
   it('should contain shadow root', async () => {
     const harness = await createFixture();
 
-    expect(harness.codeFeature.shadowRoot).to.be.ok;
+    expect(harness.strikeFeature.shadowRoot).to.be.ok;
   });
 
   it('should have expected default label', async () => {
     const harness = await createFixture();
 
-    expect(harness.codeFeature.label).to.equal('code');
+    expect(harness.strikeFeature.label).to.equal('Strikethrough');
   });
 
   it('should set custom label', async () => {
-    const harness = await createFixture({ label: 'Inline Code' });
+    const harness = await createFixture({ label: 'Strike Through' });
 
-    expect(harness.codeFeature.label).to.equal('Inline Code');
-    expect(harness.button().getAttribute('aria-label')).to.equal('Inline Code');
+    expect(harness.strikeFeature.label).to.equal('Strike Through');
+    expect(harness.button().getAttribute('aria-label')).to.equal('Strike Through');
   });
 
-  it('should render code button', async () => {
+  it('should render strikethrough button', async () => {
     const harness = await createFixture();
 
     expect(harness.button()).to.exist;
   });
 
-  it('should configure code extension', async () => {
+  it('should configure strike extension', async () => {
     const harness = await createFixture();
 
-    expect(harness.codeFeature.extensions).to.have.lengthOf(1);
-    expect(harness.codeFeature.extensions[0].name).to.equal('code');
+    expect(harness.strikeFeature.extensions).to.have.lengthOf(1);
+    expect(harness.strikeFeature.extensions[0].name).to.equal('strike');
   });
 
-  it('should toggle code when button is clicked', async () => {
+  it('should toggle strike when button is clicked', async () => {
     const harness = await createFixture();
     const editor = await harness.getEditor();
 
@@ -49,12 +49,12 @@ describe('RTE Code Feature', () => {
     editor.commands.selectAll();
     await harness.waitForUpdate();
 
-    // Click button to apply code
+    // Click button to apply strikethrough
     await harness.clickButton();
 
-    // Verify code was applied
+    // Verify strikethrough was applied
     const output = editor.getHTML();
-    expect(output).to.include('<code>test text</code>');
+    expect(output).to.include('<s>test text</s>');
   });
 
   it('should disable button when editor is disabled', async () => {
@@ -69,7 +69,7 @@ describe('RTE Code Feature', () => {
     expect(harness.button().hasAttribute('disabled')).to.be.true;
   });
 
-  it('should show active state when text has code formatting', async () => {
+  it('should show active state when text has strikethrough', async () => {
     const harness = await createFixture();
     const editor = await harness.getEditor();
 
@@ -78,27 +78,27 @@ describe('RTE Code Feature', () => {
     editor.commands.selectAll();
     await harness.waitForUpdate();
 
-    // Apply code formatting
-    editor.chain().focus().toggleCode().run();
+    // Apply strikethrough formatting
+    editor.chain().focus().toggleStrike().run();
     await harness.waitForUpdate();
 
     expect(harness.button().hasAttribute('pressed')).to.be.true;
   });
 
-  it('should not show active state when text does not have code formatting', async () => {
+  it('should not show active state when text does not have strikethrough', async () => {
     const harness = await createFixture();
 
     expect(harness.button().hasAttribute('pressed')).to.be.false;
   });
 
-  it('should toggle off code when clicking active button', async () => {
+  it('should toggle off strikethrough when clicking active button', async () => {
     const harness = await createFixture();
     const editor = await harness.getEditor();
 
-    // Set content, select it, and apply code
+    // Set content, select it, and apply strikethrough
     editor.commands.setContent('<p>test</p>');
     editor.commands.selectAll();
-    editor.chain().focus().toggleCode().run();
+    editor.chain().focus().toggleStrike().run();
     await harness.waitForUpdate();
     expect(harness.button().hasAttribute('pressed')).to.be.true;
 
@@ -107,7 +107,7 @@ describe('RTE Code Feature', () => {
     expect(harness.button().hasAttribute('pressed')).to.be.false;
   });
 
-  it('should apply code to selected text', async () => {
+  it('should apply strikethrough to selected text', async () => {
     const harness = await createFixture();
     const editor = await harness.getEditor();
 
@@ -116,95 +116,86 @@ describe('RTE Code Feature', () => {
     editor.commands.selectAll();
     await harness.waitForUpdate();
 
-    // Apply code
+    // Apply strikethrough
     await harness.clickButton();
 
     const output = editor.getHTML();
-    expect(output).to.include('<code>');
+    expect(output).to.include('<s>');
     expect(output).to.include('test text');
   });
 
-  it('should remove code from selected code text', async () => {
+  it('should remove strikethrough from selected strikethrough text', async () => {
     const harness = await createFixture();
     const editor = await harness.getEditor();
 
-    // Set code content and select it
-    editor.commands.setContent('<p><code>code text</code></p>');
+    // Set strikethrough content and select it
+    editor.commands.setContent('<p><s>strikethrough text</s></p>');
     editor.commands.selectAll();
     await harness.waitForUpdate();
 
-    // Remove code
+    // Remove strikethrough
     await harness.clickButton();
 
     const output = editor.getHTML();
-    expect(output).not.to.include('<code>');
-    expect(output).to.include('code text');
+    expect(output).not.to.include('<s>');
+    expect(output).to.include('strikethrough text');
   });
 
-  it('should apply code to partial text selection', async () => {
+  it.skip('should work with keyboard shortcut Ctrl+Shift+S', async () => {
     const harness = await createFixture();
     const editor = await harness.getEditor();
 
-    // Set content and select partial text
-    editor.commands.setContent('<p>some text here</p>');
-    editor.commands.setTextSelection({ from: 6, to: 10 }); // Select "text"
-    await harness.waitForUpdate();
-
-    // Apply code
-    await harness.clickButton();
-
-    const output = editor.getHTML();
-    expect(output).to.include('<code>text</code>');
-    expect(output).to.include('some');
-    expect(output).to.include('here');
-  });
-
-  it('should preserve text when toggling code format', async () => {
-    const harness = await createFixture();
-    const editor = await harness.getEditor();
-
-    const testText = 'preserved text';
-    editor.commands.setContent(`<p>${testText}</p>`);
+    // Set content and select it
+    editor.commands.setContent('<p>test</p>');
     editor.commands.selectAll();
+    editor.commands.focus();
     await harness.waitForUpdate();
 
-    // Apply code
-    await harness.clickButton();
-    let output = editor.getHTML();
-    expect(output).to.include(testText);
-    expect(output).to.include('<code>');
+    // Get editor DOM element from content component
+    const contextElement = harness.el.shadowRoot!.querySelector('forge-rich-text-context')!;
+    const contentElement = contextElement.shadowRoot!.querySelector('forge-rich-text-content')!;
+    const editorElement = contentElement.shadowRoot!.querySelector('.ProseMirror') as HTMLElement;
 
-    // Remove code
-    await harness.clickButton();
-    output = editor.getHTML();
-    expect(output).to.include(testText);
-    expect(output).not.to.include('<code>');
+    expect(editorElement).to.exist;
+
+    // Simulate Ctrl+Shift+S
+    const event = new KeyboardEvent('keydown', {
+      key: 's',
+      ctrlKey: true,
+      shiftKey: true,
+      bubbles: true,
+      cancelable: true
+    });
+    editorElement.dispatchEvent(event);
+    await harness.waitForUpdate();
+
+    expect(harness.button().hasAttribute('pressed')).to.be.true;
   });
 });
 
-interface CodeFixtureOptions {
+interface StrikeFixtureOptions {
   label?: string;
   disabled?: boolean;
   readonly?: boolean;
 }
 
-interface CodeFixture {
+interface StrikeFixture {
   el: RichTextEditorComponent;
-  codeFeature: RichTextFeatureCodeComponent;
+  strikeFeature: RichTextFeatureStrikeComponent;
   button: () => HTMLElement;
   clickButton: () => Promise<void>;
   getEditor: () => Promise<Editor>;
   waitForUpdate: () => Promise<void>;
 }
 
-async function createFixture(options: CodeFixtureOptions = {}): Promise<CodeFixture> {
+async function createFixture(options: StrikeFixtureOptions = {}): Promise<StrikeFixture> {
   const el = await fixture<RichTextEditorComponent>(html`
     <forge-rich-text-editor ?disabled=${options.disabled} ?readonly=${options.readonly}>
-      <forge-rte-code label=${options.label || 'code'}></forge-rte-code>
+      <forge-rte-strike label=${options.label || 'Strikethrough'}></forge-rte-strike>
     </forge-rich-text-editor>
   `);
 
-  const codeFeature = el.querySelector('forge-rte-code') as RichTextFeatureCodeComponent;
+  const strikeFeature = el.querySelector('forge-rte-strike') as RichTextFeatureStrikeComponent;
   const contextComponent = el.shadowRoot!.querySelector('forge-rich-text-context')!;
 
   // Wait for editor to initialize
@@ -212,9 +203,9 @@ async function createFixture(options: CodeFixtureOptions = {}): Promise<CodeFixt
 
   return {
     el,
-    codeFeature,
+    strikeFeature,
     button: () =>
-      codeFeature.shadowRoot!.querySelector('forge-rte-tool-button')!.shadowRoot!.querySelector('forge-icon-button')!,
+      strikeFeature.shadowRoot!.querySelector('forge-rte-tool-button')!.shadowRoot!.querySelector('forge-icon-button')!,
     async clickButton() {
       this.button().click();
       await this.waitForUpdate();
@@ -228,8 +219,8 @@ async function createFixture(options: CodeFixtureOptions = {}): Promise<CodeFixt
     async waitForUpdate() {
       await el.updateComplete;
       // Manually trigger re-render on feature to update active state
-      codeFeature.requestUpdate();
-      await codeFeature.updateComplete;
+      strikeFeature.requestUpdate();
+      await strikeFeature.updateComplete;
       // Give TipTap time to process
       await new Promise(resolve => setTimeout(resolve, 100));
     }
