+++ packages/extended/src/lib/rich-text-editor/features/tests/rte-underline.test.ts
@@ -1,11 +1,11 @@
 import { expect } from '@esm-bundle/chai';
 import { fixture, html } from '@open-wc/testing';
 import type { Editor } from '@tiptap/core';
-import { RichTextEditorComponent } from '../rich-text-editor';
-import { RteUnderlineComponent } from './rte-underline';
+import { RichTextEditorComponent } from '../../rich-text-editor';
+import { RteUnderlineComponent } from '../rte-underline';
 
-import '../rich-text-editor';
-import './rte-underline';
+import '../../rich-text-editor';
+import '../rte-underline';
 
 describe('RTE Underline Feature', () => {
   it('should contain shadow root', async () => {
