+++ packages/extended/src/lib/rich-text-editor/rich-text-renderer.scss
@@ -1,6 +1,65 @@
-@use 'prosemirror-view/style/prosemirror.css';
 @use '@tylertech/forge/sass/typography' as forge-typography;
 @use '@tylertech/forge/sass/core/styles/spacing' as forge-spacing;
+
+// Inlined from prosemirror-view/style/prosemirror.css (transitive dep — not directly resolvable via Sass)
+.ProseMirror {
+  position: relative;
+  word-wrap: break-word;
+  white-space: pre-wrap;
+  white-space: break-spaces;
+  -webkit-font-variant-ligatures: none;
+  font-variant-ligatures: none;
+  font-feature-settings: 'liga' 0;
+
+  pre {
+    white-space: pre-wrap;
+  }
+
+  li {
+    position: relative;
+  }
+}
+
+.ProseMirror-hideselection {
+  caret-color: transparent;
+
+  *::selection {
+    background: transparent;
+  }
+
+  *::-moz-selection {
+    background: transparent;
+  }
+}
+
+.ProseMirror [draggable][contenteditable='false'] {
+  user-select: text;
+}
+
+.ProseMirror-selectednode {
+  outline: 2px solid #8cf;
+}
+
+li.ProseMirror-selectednode {
+  outline: none;
+
+  &::after {
+    content: '';
+    position: absolute;
+    left: -32px;
+    right: -2px;
+    top: -2px;
+    bottom: -2px;
+    border: 2px solid #8cf;
+    pointer-events: none;
+  }
+}
+
+img.ProseMirror-separator {
+  display: inline !important;
+  border: none !important;
+  margin: 0 !important;
+}
 
 :host {
   display: block;
